+++ derive_builder_core/src/build_method.rs
@@ -1,5 +1,6 @@
 use quote::{Tokens, ToTokens};
 use syn;
+use Block;
 use BuilderPattern;
 use Initializer;
 use doc_comment::doc_comment_from;
@@ -51,7 +52,7 @@ pub struct BuildMethod<'a> {
     /// Whether the generated code should comply with `#![no_std]`.
     pub no_std: bool,
     /// Whether the generated method should use `Default::default()`.
-    pub use_default: bool,
+    pub struct_default: Option<Block>,
 }
 
 impl<'a> ToTokens for BuildMethod<'a> {
@@ -67,8 +68,8 @@ impl<'a> ToTokens for BuildMethod<'a> {
             BuilderPattern::Immutable => quote!(&self),
         };
         let doc_comment = &self.doc_comment;
-        let default = if self.use_default { 
-            quote!(let __default : #target_ty = ::std::default::Default::default();)
+        let default = if let Some(ref expr) = self.struct_default { 
+            quote!(let __default : #target_ty = #expr;)
         } else {
             quote!()
         };
@@ -132,7 +133,7 @@ macro_rules! default_build_method {
             initializers: vec![quote!(foo: self.foo,)],
             doc_comment: None,
             no_std: false,
-            use_default: false,
+            struct_default: None,
         }
     }
 }